cleaned up useless comps, renamed images to order properly in schema
